Finish Authorization Part

+++ app/Http/Controllers/Backend/RoleController.php
@@ -7,6 +7,7 @@ use App\Models\Module;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 
 class RoleController extends Controller
@@ -18,6 +19,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.roles.index');
         $roles = Role::all();
         return view('backend.roles.index', compact('roles'));
     }
@@ -30,6 +32,7 @@ class RoleController extends Controller
     public function create()
     {
         //
+        Gate::authorize('app.roles.create');
         $modules = Module::all();
         return view('backend.roles.form',compact('modules'));
     }
@@ -43,6 +46,7 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         //
+        Gate::authorize('app.roles.create');
         $this->validate($request,[
             'name'=>'required|unique:roles',
             'permissions'=>'required|array',
@@ -77,6 +81,7 @@ class RoleController extends Controller
     public function edit(Role $role)
     {
         //
+        Gate::authorize('app.roles.edit');
         $modules = Module::all();
         return view('backend.roles.form',compact('modules','role'));
     }
@@ -91,6 +96,7 @@ class RoleController extends Controller
     public function update(Request $request, Role $role)
     {
         //
+        Gate::authorize('app.roles.edit');
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name)
@@ -109,6 +115,7 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         //
+        Gate::authorize('app.roles.destroy');
         if ($role->deletable) {
             $role->delete();
             notify()->success("Role Deleted","Success");
